Crate ic_certification

IC Certification

Re-exports

pub use crate::hash_tree::*;

pub use crate::rb_tree::*;

pub use crate::nested_rb_tree::*;

Modules

certificate

hash_tree

cf https://internetcomputer.org/docs/current/references/ic-interface-spec/#certification-encoding

nested_rb_tree

rb_tree

Enums

LookupResult

A result of looking up for a certificate.

Functions

empty

Create an empty hash tree.

fork

Create a forked tree from two trees or node.

labeled

Create a labeled hash tree.

leaf

Create a leaf in the tree.

pruned

Create a pruned tree node.

pruned_from_hex

Create a pruned tree node, from a hex representation of the data. Useful for testing or hard coded values.

Type Aliases

Certificate

A Certificate as defined in https://internetcomputer.org/docs/current/references/ic-interface-spec/#certificate

Delegation

A Delegation as defined in https://internetcomputer.org/docs/current/references/ic-interface-spec/#certification-delegation

HashTree

A HashTree representing a full tree.

HashTreeNode

A HashTreeNode representing a node in a tree.

Label

For labeled HashTreeNode

SubtreeLookupResult

A result of looking up for a subtree.