ic_certification
Module hash_tree

cf https://internetcomputer.org/docs/current/references/ic-interface-spec/#certification-encoding

Structs

HashTree

A HashTree representing a full tree.

Label

For labeled HashTreeNode

Enums

HashTreeNode

A Node in the HashTree.

LookupResult

A result of looking up for a certificate.

SubtreeLookupResult

A result of looking up for a subtree.

Functions

empty

Create an empty hash tree.

fork

Create a forked tree from two trees or node.

fork_hash

Identifiably hashes a fork in the branch. Used for hashing HashTreeNode::Fork.

label

Create a labeled hash tree.

labeled_hash

Identifiably hashes a label for this branch. Used for hashing HashTreeNode::Labeled.

leaf

Create a leaf in the tree.

leaf_hash

Identifiably hashes a leaf node’s data. Used for hashing HashTreeNode::Leaf.

pruned

Create a pruned tree node.

pruned_from_hex

Create a pruned tree node, from a hex representation of the data. Useful for testing or hard coded values.

Type Aliases

Hash

Sha256 Digest: 32 bytes